Making use of separate files for each screen appbars, since all are custom
