+++ app/src/main/java/com/developersbreach/composeactors/ui/home/AppBar.kt
@@ -1,70 +1,30 @@
-package com.developersbreach.composeactors.ui.components
+package com.developersbreach.composeactors.ui.home
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
 import androidx.compose.material.Icon
-import androidx.compose.material.IconButton
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Search
-import androidx.compose.material.icons.rounded.ArrowBack
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.alpha
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.developersbreach.composeactors.R
-import com.developersbreach.composeactors.ui.details.DetailScreen
-import com.developersbreach.composeactors.ui.home.HomeScreen
 import com.developersbreach.composeactors.ui.theme.ComposeActorsTheme
 
-/**
- * @param navigateUp navigates back to previous screen.
- * @param title actor name in center of app bar.
- * AppBar for [DetailScreen]
- */
-@Composable
-fun DetailAppBar(
-    navigateUp: () -> Unit,
-    title: String
-) {
-    Box(
-        modifier = Modifier.fillMaxWidth()
-    ) {
-        IconButton(
-            onClick = navigateUp,
-            modifier = Modifier.align(alignment = Alignment.CenterStart)
-        ) {
-            Icon(
-                imageVector = Icons.Rounded.ArrowBack,
-                tint = MaterialTheme.colors.onBackground,
-                contentDescription = stringResource(id = R.string.cd_up_button)
-            )
-        }
-
-        Text(
-            text = title,
-            color = MaterialTheme.colors.onBackground,
-            style = MaterialTheme.typography.h6,
-            textAlign = TextAlign.Center,
-            modifier = Modifier
-                .fillMaxWidth()
-                .align(alignment = Alignment.Center)
-        )
-    }
-}
 
 /**
  * AppBar for [HomeScreen]
  */
 @Composable
-fun SearchBar(
+fun HomeAppBar(
     navigateToSearch: () -> Unit
 ) {
     Row(
@@ -100,6 +60,6 @@ fun SearchBar(
 @Composable
 fun SearchBarPreview() {
     ComposeActorsTheme(darkTheme = true) {
-        SearchBar { }
+        HomeAppBar { }
     }
 }
